+++ src/ndata.h
@@ -80,10 +80,10 @@ const char* ndata_name (void);
  * Individual file functions.
  */
 int ndata_exists( const char* filename );
-void* ndata_read( const char* filename, uint32_t *filesize );
-char** ndata_list( const char *path, uint32_t* nfiles );
-char** ndata_listRecursive( const char *path, uint32_t* nfiles );
-void ndata_sortName( char **files, uint32_t nfiles );
+void* ndata_read( const char* filename, size_t *filesize );
+char** ndata_list( const char *path, size_t* nfiles );
+char** ndata_listRecursive( const char *path, size_t* nfiles );
+void ndata_sortName( char **files, size_t nfiles );
 
 
 /*
